Extract common button styles in PixButtonBase
