+++ addon/styles/_pix-button-base.scss
@@ -1,0 +1,137 @@
+.pix-button {
+  @import 'reset-css';
+
+  color: $white;
+  font-family: $font-roboto;
+  font-size: 0.875rem;
+  font-weight: 500;
+  white-space: nowrap;
+  letter-spacing: .028rem;
+  cursor: pointer;
+  background-color: $blue;
+  border: 2px solid transparent;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  text-decoration: none;
+
+  &--shape-rounded {
+    border-radius: 25px;
+  }
+
+  &--shape-squircle {
+    border-radius: 4px;
+  }
+
+  &--size-big {
+    padding: 12px 24px;
+  }
+
+  &--size-small.pix-button--shape-rounded {
+    padding: 10px 24px;
+  }
+
+  &--size-small.pix-button--shape-squircle {
+    padding: 10px 16px;
+  }
+
+  &--disabled {
+    opacity: .5;
+    cursor: not-allowed;
+  }
+
+  @mixin colorizeBackground($backgroundColor, $outlineColor) {
+    background-color: $backgroundColor;
+    border-color: $backgroundColor;
+
+    &:not(.pix-button--disabled) {
+      &:hover, &:focus, &:focus-visible {
+        background-color: darken($backgroundColor, 8%);
+        box-shadow: 0 0 0 2px darken($outlineColor, 8%);
+        border-color: $white;
+        outline: none;
+      }
+
+      &:active {
+        background-color: darken($backgroundColor, 12%);
+      }
+    }
+  }
+
+  &--background-blue {
+    @include colorizeBackground($blue, $blue);
+  }
+
+  &--background-green {
+    @include colorizeBackground($green, $green);
+  }
+
+  &--background-yellow {
+    color: $grey-100;
+    @include colorizeBackground($yellow, $yellow);
+  }
+
+  &--background-red {
+    color: $white;
+    @include colorizeBackground($red, $red);
+  }
+
+  &--background-grey {
+    @include colorizeBackground($blue-zodiac, $blue-zodiac);
+  }
+
+  /* deprecated in favor of --background-transparent-light + --border */
+  &--background-transparent {
+    background-color: transparent;
+    color: $grey-50;
+    border: 2px solid $grey-50;
+  }
+
+  &--background-transparent-light {
+    background-color: transparent;
+    border-color: transparent;
+    color: $grey-90;
+
+    &:not(.pix-button--disabled) {
+      &:hover, &:focus, &:focus-visible {
+        background-color: rgba($black, 0.1);
+        outline: none;
+      }
+
+      &:active {
+        background-color: rgba($black, 0.2);
+      }
+    }
+
+    &.pix-button--border {
+      border-color: $grey-50;
+
+      &:not(.pix-button--disabled) {
+        &:hover, &:active {
+          border-color: $grey-80;
+        }
+      }
+    }
+  }
+
+  &--background-transparent-dark {
+    background-color: transparent;
+    border-color: transparent;
+    color: $white;
+
+    &:not(.pix-button--disabled) {
+      &:hover, &:focus, &:focus-visible {
+        background-color: rgba($black, 0.1);
+        outline: none;
+      }
+
+      &:active {
+        background-color: rgba($black, 0.2);
+      }
+    }
+
+    &.pix-button--border {
+      border-color: $white;
+    }
+  }
+}
