+++ addon/styles/addon.scss
@@ -4,6 +4,7 @@
 @import 'pix-background-header';
 @import 'pix-banner';
 @import 'pix-block';
+@import 'pix-button-base';
 @import 'pix-button';
 @import 'pix-collapsible';
 @import 'pix-filter-banner';
